+++ sjf4j/src/main/java/org/sjf4j/facade/fastjson2/Fastjson2StreamingIO.java
@@ -253,7 +253,7 @@ public class Fastjson2StreamingIO {
                 }
             }
             if (pi.isJojo) {
-                ((JsonObject) pojo).setDynamicMap(dynamicMap);
+                ((JsonObject) pojo)._dynamicMap(dynamicMap);
             }
             return pojo;
         }
@@ -737,7 +737,7 @@ public class Fastjson2StreamingIO {
             }
         }
         if (pi.isJojo && pi.writeDynamic) {
-            Map<String, Object> dynamicMap = ((JsonObject) node).getDynamicMap();
+            Map<String, Object> dynamicMap = ((JsonObject) node)._dynamicMap();
             if (dynamicMap != null) {
                 for (Map.Entry<String, Object> entry : dynamicMap.entrySet()) {
                     Object value = entry.getValue();