+++ sjf4j/src/main/java/org/sjf4j/facade/jackson2/Jackson2StreamingIO.java
@@ -318,7 +318,7 @@ public class Jackson2StreamingIO {
             }
             parser.nextToken();
             if (pi.isJojo) {
-                ((JsonObject) pojo).setDynamicMap(dynamicMap);
+                ((JsonObject) pojo)._dynamicMap(dynamicMap);
             }
             return pojo;
         }
@@ -879,7 +879,7 @@ public class Jackson2StreamingIO {
             }
         }
         if (pi.isJojo && pi.writeDynamic) {
-            Map<String, Object> dynamicMap = ((JsonObject) node).getDynamicMap();
+            Map<String, Object> dynamicMap = ((JsonObject) node)._dynamicMap();
             if (dynamicMap != null) {
                 for (Map.Entry<String, Object> entry : dynamicMap.entrySet()) {
                     Object value = entry.getValue();