+++ sjf4j/src/main/java/org/sjf4j/util/Loggers.java
@@ -3,4 +3,4 @@ package org.sjf4j.util;
 /**
  * Placeholder for logging helpers (currently empty).
  */
-public class LoggerUtil {}
+public class Loggers {}